requirements.txt
openai>=1.0.0
qdrant-client>=1.7.0
python-frontmatter>=1.1.0
psycopg2-binary>=2.9.0
fastapi>=0.110.0
uvicorn>=0.27.0
sentence-transformers>=2.6.0
pydantic>=2.0.0
